Fix check for granularity of a booking.
* inc/bb/bookingentry.php (BookingEntry::_legalSlot): this check ensures that
when bookings are made manually (instead of through the dropdown menu), their
time is correct, i.e., when bookings must be of 1h long, there is no attempt to
make a 1h30m booking. We must clone the SimpleDate object properly otherwise
when we addTime() to $stoptime, we also add it to $starttime, and the difference
between them will be zero which is always valid.