Create 'templates' directory and 'bumblebee' template from current theme.
With the purpose of allowing alternative themes without modifying the original
source, all theme files will now create the default bumblebee template. A
new directory, 'templates', will allow for multiple templates to exist side
by side, with choice of template made at configuration level. The actual code
and configuration for this still needs to be implemented on a separate commit.